Splitting up tests and modifying superagent to use then() instead of end() callbacks
